(engine): add initial renderer implementation

+++ engine/portal/engine/strings/string_id.cpp
@@ -11,7 +11,7 @@
 namespace portal
 {
 
-StringId::StringId(const uint128_t id): id(id)
+StringId::StringId(const uint64_t id): id(id)
 {
     string = StringRegistry::find(id);
     if (string == INVALID_STRING_VIEW)
@@ -20,12 +20,12 @@ StringId::StringId(const uint128_t id): id(id)
     }
 }
 
-StringId::StringId(const uint128_t id, const std::string_view string): id(id)
+StringId::StringId(const uint64_t id, const std::string_view string): id(id)
 {
     this->string = StringRegistry::store(id, string);
 }
 
-StringId::StringId(const uint128_t id, const std::string& string): StringId(id, std::string_view(string)) {}
+StringId::StringId(const uint64_t id, const std::string& string): StringId(id, std::string_view(string)) {}
 
 bool StringId::operator==(const StringId& other) const
 {
